Added example files
These are example files that show the basics of how to interact with flask. There are a few examples for the API, and one for the views. These are of a reasonable format to base anything on.  Just make sure to adjust names.  For example, instead of 'insta485', our app is named 'MDE'.